Clean up --src and --file handling
Fix a problem where the argument from --src always tries to use a
".bak.gz" extension, even if the argument specifies an extension.

Resolve this by accepting the arguments as-is if they have a
recognized extension.  The default for the src will have a ".bak.gz"
extension if copied from the file.  In other words:

If no arguments are specified, the default names are used:
  $ rsure signoff
  src : 2sure.bak.gz
  file: 2sure.dat.gz
  tmp : 2sure.tmp.gz

Specifying just a file (without an extension) uses that as the base:
  $ rsure -f /path/foo signoff
  src : /path/foo.bak.gz
  file: /path/foo.dat.gz
  tmp : /path/foo.tmp.gz

If the file specified has an extension (either ".bak.gz", or
".dat.gz") that will be used for just the file.  This means, the you
probably want to specify '--src' as well, otherwise the two files are
the same.

This isn't perfect, but at least catches some common use cases better,
such as a bunch of ".dat.gz" files with timestamps, and now --src and
--file can be used to compare arbitrary of these files.

[Bumps to version 0.2.1]