Provide an 'update' function
Move the duplicated logic that is in two places in main into a
function.  This will also help users of the library that also have to
duplicate this code.

However, until the feature default_type_parameter_fallback comes into
stable, rustc will have some difficulty inferring the type of this
function when called with `None` for the `src` argument.  As a
workaround, also provide the `no_path()` function that returns a
`None` that is appropriately typed.  See
https://users.rust-lang.org/t/asref-in-option-is-awkward-to-use/2596
for details.