bk: Add bkimport command
This takes a `--src <dir>` and a `--dest <dir>` argument, and loads
all of the surefiles in the source and commits them in the bkdir named
by dest (which should have already been created with bknew).

All names that match the pattern  *-*.dat.gz will be matched.  The
part of the filename before the '-' will be used as the sure 'file'
name, and the part after the hyphen will be used as the 'name' which
names the particular version.