bk: Force delta even with no changes
Since there may be multiple surefiles with no change, force a new
revision even if there are no changes.  Otherwise, these appear to be
missing, and a future import will pull them in.  It also definitively
states that the snapshot was made at that point, and there was nothing
changed about it.