Eliminate unsafe NOATIME
Rust 1.10 stabilizes the `custom_flags` method in `OpenOptionsExt`.
With this, it is no longer necessary to do an unsafe cast of the open
options to set the NOATIME flag.

With this patch, we do now require Rust 1.10 to compile.  It is
possible to revert this patch to use an earlier version.