weave: Implement writing of deltas
With this, we are able to generate deltas, and verify that they always
parse correctly.  I've also done a comparison with some large data sets
that the results are the same as what SCCS does (which isn't guaranteed,
since diff can give different results.  It works here because sccs is
using the same 'diff' program the same way).

Missing is the recording of any information about each delta in the
weave file.  As such, the client has to know the delta numbers, and keep
track of them.