weave: Parameterize the test results better
Put some constants at the top of weave/tests/sccs.rs to adjust how
extensive the test is.  The number of iterations and the size of the
files tested can both be increased.  There is also an option to disable
verification of all previous deltas.  Since the verification of previous
deltas is O(n^2), this can be prohibitive to run.

It is also useful to run with KEEPTEMP=1 in the environment, and to
compare s.tfile and sample.wave with each other.  The headers will
differ, because we don't use the same format, but the body should be the
same, as long as the local sccs uses the same diff algorithm.