weave: Put tags in weave files
Store the tags in a header on the weave files.  This allows us to
compute the next delta by looking at the existing ones.

Moving away from the sccs format, the header is stored as a single line
in the weave file, which is a json-encoded structure describing all of
the deltas in the file.