Add 'list' and '--version'
The 'list' command will show the various versions stored in a given
sure store.  This is only currently implemented for the weave store
type.  It doesn't really make sense on the plain type, and is not
implemented in the bk type, as this type is deprecated.

The '--version' argument can be given to the 'check' command to check
against a version other than the most recent.