Coordinate logging with progress meter
Allow the rsure library to install its own logger (currently based on
the env_logger, but this could be pluggable).  This logger wraps another
logger, but coordinates messages with the progress meter to print both
nicely.  It uses a lazy static mutex, so should be safe to use from as
many threads as desired.