Make hash update parallel
The crossbeam library has two significant features for this: 1. scoped
threads, which make this easier to express, and 2. channels that support
multiple receivers.  This allows the workers to just all pull from the
same channel of work.

The hashes are now delivered unordered (since smaller files can hash
faster), and we rely on the temporary sqlite3 database to return the
results in order afterwards.

This makes hash updates significantly faster in many situations.