weave: Bound memory usage
Using unbounded channels for loading from weave files results in
effectively unbounded memory usage.  Change these to bounded channels so
that memory usage stays bounded (and small).

Without this change, the channel tended to run ahead, and allocate space
for the entire file in memory before processing any of it.

This does hurt performance a bit, since it introduces a lot more context
switches.  A better fix would probably be to make the weave loader
"pull" instead of "push", which would eliminate the need for this thread
entirely.  Alternatively, we can wait until Rust has generators, and let
the compiler do the CPS conversion of the code.