Fix duplicated names in some comparison message
The node comparator is passed the fully assembled name for the level it
is at.  It then adds the final path component again, which results in
duplicated pathnames, e.g.

     [mtime,sha1          ] "./usr/bin/emacs/emacs"

instead of just "./usr/bin/emacs".  Fix this by just using the build
path, and not adding the name again.

Fixes #6